pyproject.toml
[project]
name = "multi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.12"
dependencies = [
    "a2a-sdk[http-server]>=0.3.11",
    "a2a[sdk]>=0.44",
    "asyncio>=4.0.0",
    "click>=8.3.0",
    "crewai[a2a]>=1.4.1",
    "fastapi>=0.121.1",
    "google-adk>=1.18.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "langchain-core>=1.0.4",
    "langgraph>=1.0.3",
    "python-dotenv>=1.2.1",
    "starlette>=0.49.3",
    "typing-extensions>=4.15.0",
    "uvicorn[standard]>=0.38.0",
]
